+++ src/routes/index.mdx
@@ -1,8 +1,9 @@
-import { SolidDocsLogo, Sections } from "../components/Logo.tsx";
+import { SolidDocsLogo } from "../components/Logo.tsx";
+import { SectionCards } from "../components/SectionCards.tsx";
 
 <SolidDocsLogo />
 
-<Sections />
+<SectionCards />
 
 export default function Layout(props) {
   return <div class="p-2">{props.children}</div>;
